add order for note and sort note in recylerview

+++ Acornote_Kotlin/app/src/main/java/tonnysunm/com/acornote/model/Note.kt
@@ -19,7 +19,7 @@ data class Note(
     var id: Long = 0,
 
     @ColumnInfo(index = true)
-    var order: Double,
+    var order: Long,
 
     var title: String,
